+++ src/main/java/ru/filippov/neat/dto/TokenDto.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class TokenDto {
     private final String accessToken;
     private final String refreshToken;
-    private final int accessTokenExpiredIn;
-    private final int refreshTokenExpiredIn;
+    private final long accessTokenExpiredIn;
+    private final long refreshTokenExpiredIn;
 }
